Deactivate Maven 3.1.1 in Back2BackCompatibilityTest
Maven 3.1.1 works, but is so old it wants to download from Maven Central
via HTTP instead of HTTPS. But HTTP download has been deactivated by
Sonatype a while ago. If someone works with a proxy or behind a Nexus +
firewall in an enterprise setting on an old Maven version, this can
still work. Maven 3.1.1 can use this plugin, it is simply a download
problem.